Move the rest of the sources
