+++ esp-hal/src/rtc_cntl/sleep/esp32s2.rs
@@ -1,8 +1,7 @@
-use super::{UlpWakeupSource, WakeSource, WakeTriggers, WakeupLevel};
+use super::WakeTriggers;
 use crate::{
-    gpio::{RtcFunction, RtcPin},
-    peripherals::{EXTMEM, LPWR, RTC_IO, SPI0, SPI1, SYSTEM},
-    rtc_cntl::{Rtc, WakeupSource, sleep::RtcioWakeupSource},
+    peripherals::{EXTMEM, LPWR, SPI0, SPI1, SYSTEM},
+    rtc_cntl::Rtc,
     soc::regi2c,
 };
 
@@ -65,88 +64,6 @@ pub const DG_PERI_WAIT_CYCLES: u16 = OTHER_BLOCKS_WAIT;
 pub const RTC_MEM_POWERUP_CYCLES: u8 = OTHER_BLOCKS_POWERUP;
 /// RTC memory wait cycles.
 pub const RTC_MEM_WAIT_CYCLES: u16 = OTHER_BLOCKS_WAIT;
-
-impl WakeSource for UlpWakeupSource {
-    fn apply(
-        &self,
-        _rtc: &Rtc<'_>,
-        triggers: &mut WakeTriggers,
-        sleep_config: &mut RtcSleepConfig,
-    ) {
-        if self.wake_on_interrupt {
-            triggers.insert(WakeupSource::UlpRiscv);
-        }
-        if self.wake_on_trap {
-            triggers.insert(WakeupSource::UlpRiscvTrap);
-        }
-
-        if self.clear_interrupts_on_sleep {
-            self.clear_interrupts();
-        }
-
-        // This one needs to be false to keep the ULP timer and ULP GPIO happy!
-        // Possibly relevant issue: https://github.com/espressif/esp-idf/issues/10595
-        sleep_config.set_rtc_peri_pd_en(false);
-    }
-}
-
-impl RtcioWakeupSource<'_, '_> {
-    fn apply_pin(&self, pin: &mut dyn RtcPin, level: WakeupLevel) {
-        let rtcio = RTC_IO::regs();
-
-        pin.rtc_set_config(true, true, RtcFunction::Rtc);
-
-        rtcio.pin(pin.number() as usize).modify(|_, w| unsafe {
-            w.gpio_pin_wakeup_enable().set_bit();
-            w.gpio_pin_int_type().bits(match level {
-                WakeupLevel::Low => 4,
-                WakeupLevel::High => 5,
-            })
-        });
-    }
-}
-
-impl WakeSource for RtcioWakeupSource<'_, '_> {
-    fn apply(
-        &self,
-        _rtc: &Rtc<'_>,
-        triggers: &mut WakeTriggers,
-        sleep_config: &mut RtcSleepConfig,
-    ) {
-        let mut pins = self.pins.borrow_mut();
-
-        if pins.is_empty() {
-            return;
-        }
-
-        // don't power down RTC peripherals
-        sleep_config.set_rtc_peri_pd_en(false);
-        triggers.insert(WakeupSource::Gpio);
-
-        // Since we only use RTCIO pins, we can keep deep sleep enabled.
-        let sens = crate::peripherals::SENS::regs();
-
-        // TODO: disable clock when not in use
-        sens.sar_io_mux_conf()
-            .modify(|_, w| w.iomux_clk_gate_en().set_bit());
-
-        for (pin, level) in pins.iter_mut() {
-            self.apply_pin(*pin, *level);
-        }
-    }
-}
-
-impl Drop for RtcioWakeupSource<'_, '_> {
-    fn drop(&mut self) {
-        // should we have saved the pin configuration first?
-        // set pin back to IO_MUX (input_enable and func have no effect when pin is sent
-        // to IO_MUX)
-        let mut pins = self.pins.borrow_mut();
-        for (pin, _level) in pins.iter_mut() {
-            pin.rtc_set_config(true, false, RtcFunction::Rtc);
-        }
-    }
-}
 
 bitfield::bitfield! {
     /// Configuration for the RTC sleep behavior.
