+++ esp-hal/src/rtc_cntl/sleep/mod.rs
@@ -14,9 +14,6 @@
 //!    * `ULP (Ultra-Low Power)` wake
 //!    * `BT (Bluetooth) wake` - light sleep only
 
-#[cfg(any(esp32c2, esp32c3, esp32s2, esp32s3))]
-use core::cell::RefCell;
-
 use enumset::EnumSet;
 
 use crate::{
@@ -61,197 +58,6 @@ pub enum WakeupLevel {
     #[default]
     ///  The wake-up event is triggered when the pin is high.
     High,
-}
-
-#[procmacros::doc_replace(
-    "pin0" => {
-        cfg(any(esp32c3, esp32c2)) => "GPIO2",
-        cfg(any(esp32s2, esp32s3)) => "GPIO17"
-    },
-    "pin1" => {
-        cfg(any(esp32c3, esp32c2)) => "GPIO3",
-        cfg(any(esp32s2, esp32s3)) => "GPIO18"
-    },
-    "rtc_pin_trait" => {
-        cfg(any(esp32c3, esp32c2)) => "gpio::RtcPinWithResistors",
-        cfg(any(esp32s2, esp32s3)) => "gpio::RtcPin"
-    },
-)]
-/// RTC_IO wakeup source
-///
-/// RTC_IO wakeup allows configuring any combination of RTC_IO pins with
-/// arbitrary wakeup levels to wake up the chip from sleep. This wakeup source
-/// can be used to wake up from both light and deep sleep.
-///
-/// ```rust, no_run
-/// # {before_snippet}
-/// # use esp_hal::delay::Delay;
-/// # use esp_hal::gpio::{self, Input, InputConfig, Pull};
-/// # use esp_hal::rtc_cntl::{reset_reason,
-/// #   sleep::{LowPower, RtcioWakeupSource, TimerWakeupSource, WakeupLevel},
-/// #   wakeup_cause, SocResetReason
-/// # };
-/// # use esp_hal::system::Cpu;
-/// # use esp_hal::time::Duration;
-///
-/// let mut lpwr = LowPower::new(peripherals.LPWR);
-///
-/// let reason = reset_reason(Cpu::ProCpu);
-/// let wake_reason = wakeup_cause();
-///
-/// println!("{:?} {:?}", reason, wake_reason);
-///
-/// let delay = Delay::new();
-/// let timer = TimerWakeupSource::new(Duration::from_secs(10));
-/// let wakeup_pins: &mut [(&mut dyn __rtc_pin_trait__, WakeupLevel)] = &mut [
-///     (&mut peripherals.__pin0__, WakeupLevel::Low),
-///     (&mut peripherals.__pin1__, WakeupLevel::High),
-/// ];
-///
-/// let rtcio = RtcioWakeupSource::new(wakeup_pins);
-/// delay.delay_millis(100);
-/// lpwr.sleep_deep(&[&timer, &rtcio]);
-///
-/// # {after_snippet}
-/// ```
-#[cfg(any(esp32c3, esp32s2, esp32s3, esp32c2))]
-pub struct RtcioWakeupSource<'a, 'b> {
-    pins: RefCell<&'a mut [(&'b mut dyn RtcIoWakeupPinType, WakeupLevel)]>,
-}
-
-#[cfg(any(esp32c3, esp32s2, esp32s3, esp32c2))]
-impl<'a, 'b> RtcioWakeupSource<'a, 'b> {
-    /// Creates a new external GPIO wake-up source.
-    pub fn new(pins: &'a mut [(&'b mut dyn RtcIoWakeupPinType, WakeupLevel)]) -> Self {
-        Self {
-            pins: RefCell::new(pins),
-        }
-    }
-}
-
-/// LP Core wakeup source
-///
-/// Wake up from LP core. This wakeup source
-/// can be used to wake up from both light and deep sleep.
-#[cfg(esp32c6)]
-pub struct WakeFromLpCoreWakeupSource {}
-
-#[cfg(esp32c6)]
-impl WakeFromLpCoreWakeupSource {
-    /// Create a new instance of `WakeFromLpCoreWakeupSource`
-    pub fn new() -> Self {
-        Self {}
-    }
-}
-
-#[cfg(esp32c6)]
-impl Default for WakeFromLpCoreWakeupSource {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-/// ULP wakeup source
-///
-/// Wake up from ULP software interrupt, and/or ULP-RISCV Trap condition.
-/// Both of these triggers are enabled by default.
-/// This source will clear any outstanding software interrupts prior to entering sleep, by default.
-///
-/// S2 supports the following triggers (Refer to ESP32-S2 Technical Reference Manual, Table 9.4-3.
-/// Wakeup Source)
-///  - ULP-FSM software interrupt (unsure if this ALSO supports ULP-RISCV software interrupt)
-///  - ULP-RISCV Trap
-///
-/// S3 supports the following triggers (Refer to ESP32-S3 Technical Reference Manual, Table 10.4-3.
-/// Wakeup Source)
-///  - ULP-FSM software interrupt and ULP-RISCV software interrupt
-///  - ULP-RISCV Trap
-///
-/// This wakeup source can be used to wake up from both light and deep sleep.
-#[cfg(any(esp32s2, esp32s3))]
-pub struct UlpWakeupSource {
-    wake_on_interrupt: bool,
-    wake_on_trap: bool,
-    clear_interrupts_on_sleep: bool,
-}
-
-#[cfg(any(esp32s2, esp32s3))]
-impl UlpWakeupSource {
-    /// Create a new instance of `WakeFromUlpWakeupSource`
-    pub const fn new() -> Self {
-        Self {
-            wake_on_interrupt: true,
-            wake_on_trap: true,
-            clear_interrupts_on_sleep: true,
-        }
-    }
-
-    /// Enable wakeup triggered by software interrupt from ULP-FSM or ULP-RISCV
-    pub fn set_wake_on_interrupt(mut self, value: bool) -> Self {
-        self.wake_on_interrupt = value;
-        self
-    }
-
-    /// Enable wakeup triggered by ULP-RISCV Trap
-    pub fn set_wake_on_trap(mut self, value: bool) -> Self {
-        self.wake_on_trap = value;
-        self
-    }
-
-    /// Enable clearing of latched wake-up interrupts prior to entering sleep
-    pub fn set_clear_interrupts_on_sleep(mut self, value: bool) -> Self {
-        self.clear_interrupts_on_sleep = value;
-        self
-    }
-
-    /// Clears the wake-up interrupts
-    pub fn clear_interrupts(&self) {
-        crate::peripherals::LPWR::regs().int_clr().write(|w| {
-            w.cocpu_trap().clear_bit_by_one();
-            w.cocpu().clear_bit_by_one();
-            w.ulp_cp().clear_bit_by_one()
-        });
-    }
-}
-
-#[cfg(any(esp32s2, esp32s3))]
-impl Default for UlpWakeupSource {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-/// GPIO wakeup source
-///
-/// Wake up from GPIO high or low level. Any pin can be used with this wake up
-/// source. Configure the pin for wake up via
-/// [crate::gpio::Input::wakeup_enable].
-///
-/// This wakeup source can be used to wake up from light sleep only.
-pub struct GpioWakeupSource {}
-
-impl GpioWakeupSource {
-    /// Create a new instance of [GpioWakeupSource]
-    pub fn new() -> Self {
-        Self {}
-    }
-}
-
-impl Default for GpioWakeupSource {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl WakeSource for GpioWakeupSource {
-    fn apply(
-        &self,
-        _rtc: &Rtc<'_>,
-        triggers: &mut WakeTriggers,
-        _sleep_config: &mut RtcSleepConfig,
-    ) {
-        triggers.insert(WakeupSource::Gpio);
-    }
 }
 
 /// The set of wakeup sources configured to end a sleep.
